warning修正中

+++ ACG/ACG/ObjHero.cpp
@@ -42,18 +42,18 @@ void CObjHero::Action()
 		//右下
 		if (i == 1)
 		{
-			pos_x = HERO_SIZE_WIDTH;
+			pos_x = (int)HERO_SIZE_WIDTH;
 		}
 		//左下
 		else if(i == 2)
 		{
-			pos_x = HERO_SIZE_WIDTH / 2;
+			pos_x = (int)HERO_SIZE_WIDTH / 2;
 		}
 		//主人公のX位置(マップの要素数)
-		int x = (m_px + pos_x) / BLOCK_SIZE;
+		int x = ((int)m_px + pos_x) / (int)BLOCK_SIZE;
 		//主人公のY位置(マップの要素数)
 		//少し下にする
-		int y = (m_py + 1 + HERO_SIZE_HEIGHT) / BLOCK_SIZE;
+		int y = ((int)m_py + 1 + (int)HERO_SIZE_HEIGHT) / (int)BLOCK_SIZE;
 		//ブロック情報が0で無いなら取得
 		if (objmap->GetMap(x, y)!= 0)
 		{
